fix: use the resolved vite root
Although vite-plugin-svelte needs to infer a root in order to pre-scan
svelte dependencies, it should use the root returned in configResolved
to ensure compatibility with backend integrations and other plugins that
might modify the root.

Since vite-plugin-svelte uses `enforce: 'pre'`, it's not viable for
other plugins to workaround this by using `enforce: 'pre'`, since the
behavior would become order-dependent on how the user applies the
plugins, which is error-prone and would result in a bad user experience.