Fixes: Doesn't work with firefox on LinuxInspector
The 'meta-shift' shortcut won't work on firefox when used on linux.
Also in KDE the shortcut is preset to draw mouse trails on the screen.
Switching to control-shift on linux but is a little tricke to work with,
because of the way it handles the click to activate the browser window again.

Having tried all sort of options/combinations it seems the best working for all browsers
is the combination of "control-shift" and holdMode: true